+++ arlulacore/exception.py
@@ -1,0 +1,10 @@
+'''
+    Custom Exception Class
+'''
+
+class ArlulaSessionError(Exception):
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return self.value
